+++ wasatch/InGaAsCorrection.py
@@ -16,7 +16,7 @@ class InGaAsCorrection:
         self.offsets = None 
         self.slopes = None 
 
-    def to_json(self): 
+    def to_dict(self): 
         return vars(self)
 
     def parse_json_data(self, data):