+++ wasatch/XSAccessoryConnector.py
@@ -59,7 +59,7 @@ class XSAccState:
     def __repr__(self):
         return f"XSAccState < gpio enabled {self.gpio_enabled}, 5V enabled {self.acc_5V_enabled}, 5V good {self.acc_5V_good} >"
 
-    def to_json(self): 
+    def to_dict(self): 
         return vars(self)
 
 class XSGPIOState:
@@ -149,7 +149,7 @@ class XSGPIOState:
     def __repr__(self):
         return f"XSGPIOState < num {self.num}, control {self.get_control_str()}, direction {self.get_direction_str()}, value {self.get_value_str()}, func {self.get_function_str()} >"
 
-    def to_json(self): 
+    def to_dict(self): 
         return vars(self)
 
 class XSContinuousStrobe:
@@ -163,7 +163,7 @@ class XSContinuousStrobe:
     def __repr__(self):
         return f"XSContinuousStrobe < period {self.period_us}us, width {self.width_us}us, delay {self.delay_us}us, repeat {self.repeat_count} >"
 
-    def to_json(self): 
+    def to_dict(self): 
         return vars(self)
 
 class XSAccessoryConnector:
@@ -184,7 +184,7 @@ class XSAccessoryConnector:
     def __repr__(self):
         return f"XSAccessoryConnector < acc_state {self.acc_state}, GPIO1 {self.state_gpio1}, GPIO2 {self.state_gpio2}, strobe {self.cont_strobe} >"
 
-    def to_json(self): 
+    def to_dict(self): 
         return vars(self)
 
     def dump(self):